requirements.txt
streamlit
langgraph
langchain
langchain-openai
python-dotenv
sentence-transformers
faiss-cpu
numpy
pytest
pyyaml
pypdf
python-docx